refactor(view): split model.rs into model/{mod,cards,boards,board_sort}.rs
model.rs was 909 lines, over CLAUDE.md's <300-line module guideline,
flagged by the split-opus review as pre-existing debt worth splitting
at the extraction point. Split along the natural concern boundary:

- model/mod.rs: struct definition, Default impl, and the shared
  lifecycle surface (load_from_snapshot, the private
  rebuild_displayed_partitions/sort_partitions helpers, columns/
  sprints/graph accessors)
- model/cards.rs: card-focused accessors (all_cards, card_by_id,
  displayed_cards, live_cards, archived_cards, archived_card_markers,
  archived_card_ids)
- model/boards.rs: plain board accessors (boards, live_boards,
  archived_boards_view, archived_boards, archived_board_ids,
  displayed_boards, board_by_id)
- model/boards_sort.rs: board sort state and mutation (board_sort,
  set_board_sort, toggle_board_sort_order, set_board_sort_from_config)
  plus its tests, which dominated the original file's line count

Pure structural move -- no behavior change, no public API change.
Child modules access Model's private fields via Rust's
descendant-module visibility rule, so no field was widened to
pub(crate). All 25 pre-existing model tests moved verbatim and pass
unchanged; kanban-tui needs no changes since it only ever referenced
Model by its (unchanged) public method signatures.
